Create hoc 'withResizing' to resize collapsible Menu based on height available
